+++ src/main/java/Duke.java
@@ -1,3 +1,10 @@
+import commands.Command;
+import dukeexception.DukeException;
+import parser.Parser;
+import storage.Storage;
+import tasklist.TaskList;
+import ui.Ui;
+
 public class Duke {
 
     private Storage storage;
