+++ src/main/java/commands/Command.java
@@ -1,3 +1,10 @@
+package commands;
+
+import dukeexception.DukeException;
+import storage.Storage;
+import tasklist.TaskList;
+import ui.Ui;
+
 public abstract class Command {
 
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException;
