+++ src/main/java/commands/ExitCommand.java
@@ -1,3 +1,9 @@
+package commands;
+
+import storage.Storage;
+import tasklist.TaskList;
+import ui.Ui;
+
 public class ExitCommand extends Command {
 
     @Override
