+++ src/main/java/commands/SearchCommand.java
@@ -1,3 +1,12 @@
+package commands;
+
+import storage.Storage;
+import task.Deadline;
+import task.Event;
+import task.Task;
+import tasklist.TaskList;
+import ui.Ui;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
