+++ src/main/java/commands/UnknownCommand.java
@@ -1,3 +1,9 @@
+package commands;
+
+import storage.Storage;
+import tasklist.TaskList;
+import ui.Ui;
+
 public class UnknownCommand extends Command {
 
     @Override
