+++ src/main/java/dukeexception/DukeException.java
@@ -1,3 +1,5 @@
+package dukeexception;
+
 public class DukeException extends Exception {
 
     public DukeException() { }
