+++ src/main/java/dukeexception/WrongInputException.java
@@ -1,3 +1,5 @@
+package dukeexception;
+
 public class WrongInputException extends DukeException {
 
     public String input;
